remove re_types from datadependencies again by using example types; reduces diff quite a bit as well!
