+++ crates/re_arrow_store/Cargo.toml
@@ -20,7 +20,7 @@ all-features = true
 default = []
 
 ## Enables access to re_types' test components/datatypes.
-testing = ["re_types/testing"]
+testing = ["re_types/testing", "re_log_types/testing"]
 
 ## Enables `parking_lot`'s deadlock detection background thread.
 deadlock_detection = ["parking_lot/deadlock_detection"]
