+++ crates/re_log_types/src/data_table.rs
@@ -1248,3 +1248,77 @@ impl DataTable {
         Ok(())
     }
 }
+
+// ---
+
+/// Crafts a simple but interesting [`DataTable`].
+#[cfg(not(target_arch = "wasm32"))]
+#[cfg(feature = "testing")]
+impl DataTable {
+    pub fn example(timeless: bool) -> DataTable {
+        use crate::{
+            example_components::{MyColor, MyLabel, MyPoint},
+            Time,
+        };
+
+        let table_id = TableId::random();
+
+        let mut tick = 0i64;
+        let mut timepoint = |frame_nr: i64| {
+            let tp = if timeless {
+                TimePoint::timeless()
+            } else {
+                TimePoint::from([
+                    (Timeline::log_time(), Time::now().into()),
+                    (Timeline::log_tick(), tick.into()),
+                    (Timeline::new_sequence("frame_nr"), frame_nr.into()),
+                ])
+            };
+            tick += 1;
+            tp
+        };
+
+        let row0 = {
+            let num_instances = 2;
+            let positions: &[MyPoint] = &[MyPoint::new(10.0, 10.0), MyPoint::new(20.0, 20.0)];
+            let colors: &[_] = &[MyColor(0x8080_80FF)];
+            let labels: &[MyLabel] = &[];
+
+            DataRow::from_cells3(
+                RowId::random(),
+                "a",
+                timepoint(1),
+                num_instances,
+                (positions, colors, labels),
+            )
+            .unwrap()
+        };
+
+        let row1 = {
+            let num_instances = 0;
+            let colors: &[MyColor] = &[];
+
+            DataRow::from_cells1(RowId::random(), "b", timepoint(1), num_instances, colors).unwrap()
+        };
+
+        let row2 = {
+            let num_instances = 1;
+            let colors: &[_] = &[MyColor(0xFFFF_FFFF)];
+            let labels: &[_] = &[MyLabel("hey".into())];
+
+            DataRow::from_cells2(
+                RowId::random(),
+                "c",
+                timepoint(2),
+                num_instances,
+                (colors, labels),
+            )
+            .unwrap()
+        };
+
+        let mut table = DataTable::from_rows(table_id, [row0, row1, row2]);
+        table.compute_all_size_bytes();
+
+        table
+    }
+}
