+++ crates/re_sdk/src/recording_stream.rs
@@ -1589,7 +1589,6 @@ impl RecordingStream {
 
 #[cfg(test)]
 mod tests {
-    use re_arrow_store::example_datatable;
     use re_log_types::{DataTable, RowId};
 
     use super::*;
@@ -1610,7 +1609,7 @@ mod tests {
 
         let store_info = rec.store_info().cloned().unwrap();
 
-        let mut table = example_datatable(false);
+        let mut table = DataTable::example(false);
         table.compute_all_size_bytes();
         for row in table.to_rows() {
             rec.record_row(row.unwrap(), false);
@@ -1677,7 +1676,7 @@ mod tests {
 
         let store_info = rec.store_info().cloned().unwrap();
 
-        let mut table = example_datatable(false);
+        let mut table = DataTable::example(false);
         table.compute_all_size_bytes();
         for row in table.to_rows() {
             rec.record_row(row.unwrap(), false);
@@ -1757,7 +1756,7 @@ mod tests {
 
         let store_info = rec.store_info().cloned().unwrap();
 
-        let mut table = example_datatable(false);
+        let mut table = DataTable::example(false);
         table.compute_all_size_bytes();
         for row in table.to_rows() {
             rec.record_row(row.unwrap(), false);
@@ -1811,7 +1810,7 @@ mod tests {
             .memory()
             .unwrap();
 
-        let mut table = example_datatable(false);
+        let mut table = DataTable::example(false);
         table.compute_all_size_bytes();
         for row in table.to_rows() {
             rec.record_row(row.unwrap(), false);
